BistRunResult: dynamically compute files_with_no_tests
This makes BistRunResults correctly additive for multiple test runs on the same files. files_with_no_tests is not additive if you combine these: it's subtractive; it's the set difference of the additive results of the processed files and the files that had tests. This change makes it more likely that the RunResults object could be used for combining test results from multiple test interpreters, such as both Testify/Octave BISTs and doctest.